requirements.txt
aiogram>=3.3.0
python-dotenv>=1.0.0
supabase>=2.3.4
hurry.filesize
fastapi
uvicorn
